Improve frontend private asset handling
* Define new frontend_use_restricted_assets variable (boolean).
    * If True, copy private key; fetch private asset gem; build as normal; delete key.
    * If False, just build the frontend without the private asset gem.
* See dpla/frontend#45 for related frontend changes.

+++ ansible/roles/frontend/files/build_frontend.sh
@@ -1,25 +1,30 @@
 #!/bin/bash
 
 USE_VERSION=$1
+USE_ASSETS=${2:-unrestricted}
 export PATH=$HOME/.rbenv/bin:$PATH
 
 eval "`rbenv init -`"
-# Start ssh-agent and set environment variables.
-# Work-around for private GitHub repository in Gemfile.
-eval `ssh-agent`
-ssh-add $HOME/git_private_key
 
 cd $HOME/frontend
 
 rbenv shell $USE_VERSION
 
-bundle install
-bundle update dpla_frontend_assets
+if [[ "$USE_ASSETS" = "restricted" ]]; then
+    # Start ssh-agent and set environment variables.
+    # Work-around for private GitHub repository in Gemfile.
+    eval `ssh-agent`
+    ssh-add $HOME/git_private_key
+    bundle install
+    bundle update dpla_frontend_assets
+    # Variable set above by ssh-agent
+    kill $SSH_AGENT_PID
+else
+    bundle install --without dpla_branding
+fi
+
 rbenv rehash
 bundle exec rake assets:precompile
-
-# Variable set above by ssh-agent
-kill $SSH_AGENT_PID
 
 /usr/bin/rsync -ruptolg --checksum --delete --delay-updates \
     --exclude 'log' \
